add some more documentation commentary
