+++ packages/@jsii/go-runtime/jsii-runtime-go/api/doc.go
@@ -1,0 +1,4 @@
+// Package api contains shared type definisions used by several modules part of
+// the jsii runtime for go. It helps avoid creating circular dependencies
+// between the other modules.
+package api
