+++ packages/@jsii/go-runtime/jsii-runtime-go/kernel/doc.go
@@ -1,0 +1,4 @@
+// Package kernel defines the interface for go programs to interact with the
+// @jsii/kernel node process. This module completely abstracts managament of the
+// child process.
+package kernel
