+++ packages/@jsii/go-runtime/jsii-runtime-go/typeregistry/doc.go
@@ -1,0 +1,5 @@
+// Package typeregistry offers features useful to manage the registration and
+// operation of types in the context of the jsii runtime. These are used to
+// support type conversion for values exchanged between the child node process
+// and the go application.
+package typeregistry
